Move use-git-gutter to version-control/git/controllers
